refactor: change folder structure of table components for more readability
affects: @medly-components/core
