+++ packages/core/src/components/Table/Body/Cell/Cell.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@test-utils';
 import React from 'react';
-import Checkbox from '../../Checkbox';
-import Text from '../../Text';
+import Checkbox from '../../../Checkbox';
+import Text from '../../../Text';
 import { Cell } from './Cell.styled';
 
 describe('Cell', () => {
