+++ packages/core/src/components/Table/Body/Row/types.ts
@@ -1,5 +1,5 @@
 import { HTMLProps, WithThemeProp } from '@medly-components/utils';
-import { GridTemplateProps } from '../types';
+import { GridTemplateProps } from '../../types';
 
 export type Props = GridTemplateProps &
     HTMLProps<HTMLTableRowElement> &
