+++ packages/core/src/components/Table/GroupCell/GroupCell.styled.tsx
@@ -1,6 +1,5 @@
-import { styled, WithThemeProp } from '@medly-components/utils';
+import { styled } from '@medly-components/utils';
 import Text from '../../Text';
-import Cell from '../Cell';
 import { getBorder } from '../Table.styled';
 import { Props } from './types';
 
@@ -12,20 +11,11 @@ export const GroupCell = styled('div').attrs(({ gridTemplateColumns }: Props) =>
     display: grid;
     position: relative;
     opacity: ${({ hidden }) => (hidden ? '0' : '1')};
-    width: ${({ hidden }) => (hidden ? '0px' : '100%')};
-    height: ${({ hidden }) => (hidden ? '0px' : '100%')};
-    & + &,
-    & + ${Cell.Style} {
-        &::before {
-            ${({ showBorders }) => showBorders && getBorder('left')}
-        }
-    }
+    width: ${({ hidden }) => (hidden ? '0' : '100%')};
+    height: ${({ hidden }) => (hidden ? '0' : '100%')};
 `;
-GroupCell.defaultProps = {
-    showBorders: false
-};
 
-export const GroupCellTitle = styled(Text.Style)<WithThemeProp>`
+export const GroupCellTitle = styled(Text.Style)`
     position: relative;
     text-align: center;
     padding: 5px 0 5px 10px;
