+++ packages/core/src/components/Table/GroupCell/types.ts
@@ -1,5 +1,4 @@
 import { GridTemplateProps } from '../types';
 export interface Props extends GridTemplateProps {
     hidden: boolean;
-    showBorders?: boolean;
 }
