+++ packages/core/src/components/Table/Head/HeadRow/HeadRow.styled.tsx
@@ -1,0 +1,32 @@
+import { styled } from '@medly-components/utils';
+import { GroupCell } from '../../GroupCell';
+import { getBorder } from '../../Table.styled';
+import { GridTemplateProps } from '../../types';
+import HeadCell from '../HeadCell';
+import { Props } from './types';
+
+export const HeadRow = styled('tr').attrs(({ gridTemplateColumns }: GridTemplateProps) => ({
+    style: {
+        gridTemplateColumns
+    }
+}))<Props>`
+    display: grid;
+    align-items: end;
+    user-select: none;
+    position: sticky;
+    top: 0;
+    z-index: 2;
+
+    & > * {
+        background-color: ${({ theme }) => theme.table.headerColor};
+        border-bottom: 1px solid ${({ theme }) => theme.table.borderColor};
+    }
+
+
+    ${GroupCell} + ${GroupCell},
+    ${GroupCell} + ${HeadCell.Style} {
+        &::before {
+            ${getBorder('left')}
+        }
+    }
+`;
