+++ packages/core/src/components/Table/Head/HeadRow/index.ts
@@ -1,0 +1,1 @@
+export { HeadRow as default } from './HeadRow.styled';
