+++ packages/core/src/components/Table/Head/HeadRow/types.ts
@@ -1,0 +1,4 @@
+import { HTMLProps, WithThemeProp } from '@medly-components/utils';
+import { GridTemplateProps } from '../../types';
+
+export type Props = GridTemplateProps & HTMLProps<HTMLTableRowElement> & WithThemeProp;
